+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - ref(nextjs): Clean up minor things (#258)
+- ref(nextjs): Replace old Next.js wizard (#262)
 
 ## 3.1.0